Move the old glexperiment stuff out of the way.

+++ old/polygon.pxd
@@ -13,11 +13,11 @@ cdef class Polygon:
     cdef float *data
 
     cpdef Polygon copy(Polygon self)
-    cpdef void offset(Polygon self, float x, float y, float z)
+    cpdef void offset(Polygon self, double x, double y, double z)
     cpdef void multiply_matrix(Polygon self, int offset, int size, Matrix matrix)
     cpdef void perspective_divide(Polygon self)
 
-cpdef Polygon rectangle(float w, float h, float tw, float th)
+cpdef Polygon rectangle(double w, double h, double tw, double th)
 
 cpdef intersect(Polygon a, Polygon b, int rvstride)
 cpdef barycentric(Polygon a, Polygon b, int offset)
